[FLINK-4552] Refactor WindowOperator/Trigger Tests

+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/runtime/operators/windowing/WindowOperator.java
@@ -34,7 +34,6 @@ import org.apache.flink.api.common.state.ValueStateDescriptor;
 import org.apache.flink.api.common.typeinfo.TypeInformation;
 import org.apache.flink.api.common.typeutils.TypeSerializer;
 import org.apache.flink.api.java.functions.KeySelector;
-import org.apache.flink.api.java.tuple.Tuple1;
 import org.apache.flink.api.java.tuple.Tuple2;
 import org.apache.flink.api.java.typeutils.InputTypeConfigurable;
 import org.apache.flink.api.java.typeutils.TypeExtractor;
@@ -117,6 +116,8 @@ public class WindowOperator<K, IN, ACC, OUT, W extends Window>
 
 	protected final StateDescriptor<? extends AppendingState<IN, ACC>, ?> windowStateDescriptor;
 
+	protected final ListStateDescriptor<Tuple2<W, W>> mergingWindowsDescriptor;
+
 	/**
 	 * This is used to copy the incoming element because it can be put into several window
 	 * buffers.
@@ -181,8 +182,6 @@ public class WindowOperator<K, IN, ACC, OUT, W extends Window>
 	protected transient Set<Timer<K, W>> watermarkTimers;
 	protected transient PriorityQueue<Timer<K, W>> watermarkTimersQueue;
 
-	protected transient Map<K, MergingWindowSet<W>> mergingWindowsByKey;
-
 	/**
 	 * Creates a new {@code WindowOperator} based on the given policies and user functions.
 	 */
@@ -210,6 +209,14 @@ public class WindowOperator<K, IN, ACC, OUT, W extends Window>
 		this.windowStateDescriptor = windowStateDescriptor;
 		this.trigger = checkNotNull(trigger);
 		this.allowedLateness = allowedLateness;
+
+		if (windowAssigner instanceof MergingWindowAssigner) {
+			@SuppressWarnings({"unchecked", "rawtypes"})
+			TupleSerializer<Tuple2<W, W>> tupleSerializer = new TupleSerializer<>((Class) Tuple2.class, new TypeSerializer[] {windowSerializer, windowSerializer} );
+			mergingWindowsDescriptor = new ListStateDescriptor<>("merging-window-set", tupleSerializer);
+		} else {
+			mergingWindowsDescriptor = null;
+		}
 
 		setChainingStrategy(ChainingStrategy.ALWAYS);
 	}
@@ -259,10 +266,6 @@ public class WindowOperator<K, IN, ACC, OUT, W extends Window>
 			}
 		};
 
-		if (windowAssigner instanceof MergingWindowAssigner) {
-			mergingWindowsByKey = new HashMap<>();
-		}
-
 		currentWatermark = Long.MIN_VALUE;
 	}
 
@@ -276,7 +279,6 @@ public class WindowOperator<K, IN, ACC, OUT, W extends Window>
 		processingTimeTimersQueue = null;
 		context = null;
 		windowAssignerContext = null;
-		mergingWindowsByKey = null;
 	}
 
 	@Override
@@ -289,7 +291,6 @@ public class WindowOperator<K, IN, ACC, OUT, W extends Window>
 		processingTimeTimersQueue = null;
 		context = null;
 		windowAssignerContext = null;
-		mergingWindowsByKey = null;
 	}
 
 	@Override
@@ -304,9 +305,6 @@ public class WindowOperator<K, IN, ACC, OUT, W extends Window>
 			MergingWindowSet<W> mergingWindows = getMergingWindowSet();
 
 			for (W window: elementWindows) {
-				// If there is a merge, it can only result in a window that contains our new
-				// element because we always eagerly merge
-				final Tuple1<TriggerResult> mergeTriggerResult = new Tuple1<>(TriggerResult.CONTINUE);
 
 				// adding the new window might result in a merge, in that case the actualWindow
 				// is the merged window and we work with that. If we don't merge then
@@ -319,8 +317,7 @@ public class WindowOperator<K, IN, ACC, OUT, W extends Window>
 						context.key = key;
 						context.window = mergeResult;
 
-						// store for later use
-						mergeTriggerResult.f0 = context.onMerge(mergedWindows);
+						context.onMerge(mergedWindows);
 
 						for (W m: mergedWindows) {
 							context.window = m;
@@ -348,8 +345,8 @@ public class WindowOperator<K, IN, ACC, OUT, W extends Window>
 					throw new IllegalStateException("Window " + window + " is not in in-flight window set.");
 				}
 
-				AppendingState<IN, ACC> windowState = getPartitionedState(
-					stateWindow, windowSerializer, windowStateDescriptor);
+				AppendingState<IN, ACC> windowState =
+						getPartitionedState(stateWindow, windowSerializer, windowStateDescriptor);
 				windowState.add(element.getValue());
 
 				context.key = key;
@@ -358,22 +355,23 @@ public class WindowOperator<K, IN, ACC, OUT, W extends Window>
 				// we might have already fired because of a merge but still call onElement
 				// on the (possibly merged) window
 				TriggerResult triggerResult = context.onElement(element);
-				TriggerResult combinedTriggerResult = TriggerResult.merge(triggerResult, mergeTriggerResult.f0);
 
-				if (combinedTriggerResult.isFire()) {
+				if (triggerResult.isFire()) {
 					ACC contents = windowState.get();
 					if (contents == null) {
 						continue;
 					}
-					fire(actualWindow, contents);
+					emitWindowContents(actualWindow, contents);
 				}
 
-				if (combinedTriggerResult.isPurge()) {
-					cleanup(actualWindow, windowState, mergingWindows);
-				} else {
-					registerCleanupTimer(actualWindow);
+				if (triggerResult.isPurge()) {
+					windowState.clear();
 				}
+				registerCleanupTimer(actualWindow);
 			}
+
+			// need to make sure to update the merging state in state
+			mergingWindows.persist();
 		} else {
 			for (W window: elementWindows) {
 
@@ -382,8 +380,8 @@ public class WindowOperator<K, IN, ACC, OUT, W extends Window>
 					continue;
 				}
 
-				AppendingState<IN, ACC> windowState = getPartitionedState(
-					window, windowSerializer, windowStateDescriptor);
+				AppendingState<IN, ACC> windowState =
+						getPartitionedState(window, windowSerializer, windowStateDescriptor);
 				windowState.add(element.getValue());
 
 				context.key = key;
@@ -396,16 +394,25 @@ public class WindowOperator<K, IN, ACC, OUT, W extends Window>
 					if (contents == null) {
 						continue;
 					}
-					fire(window, contents);
+					emitWindowContents(window, contents);
 				}
 
 				if (triggerResult.isPurge()) {
-					cleanup(window, windowState, null);
-				} else {
-					registerCleanupTimer(window);
+					windowState.clear();
 				}
+				registerCleanupTimer(window);
 			}
 		}
+	}
+
+	protected MergingWindowSet<W> getMergingWindowSet() throws Exception {
+		ListState<Tuple2<W, W>> mergeState =
+				getPartitionedState(VoidNamespace.INSTANCE, VoidNamespaceSerializer.INSTANCE, mergingWindowsDescriptor);
+
+		@SuppressWarnings({"unchecked", "rawtypes"})
+		MergingWindowSet<W> mergingWindows = new MergingWindowSet<>((MergingWindowAssigner) windowAssigner, mergeState);
+
+		return mergingWindows;
 	}
 
 	@Override
@@ -448,13 +455,19 @@ public class WindowOperator<K, IN, ACC, OUT, W extends Window>
 
 				TriggerResult triggerResult = context.onEventTime(timer.timestamp);
 				if (triggerResult.isFire()) {
-					fire(context.window, contents);
+					emitWindowContents(context.window, contents);
 				}
 
-				if (triggerResult.isPurge() || (windowAssigner.isEventTime() && isCleanupTime(context.window, timer.timestamp))) {
-					cleanup(context.window, windowState, mergingWindows);
+				if (windowAssigner.isEventTime() && isCleanupTime(context.window, timer.timestamp)) {
+					clearAllState(context.window, windowState, mergingWindows);
+				} else if (triggerResult.isPurge()) {
+					windowState.clear();
 				}
 
+				if (mergingWindows != null) {
+					// need to make sure to update the merging state in state
+					mergingWindows.persist();
+				}
 			} else {
 				fire = false;
 			}
@@ -510,13 +523,19 @@ public class WindowOperator<K, IN, ACC, OUT, W extends Window>
 
 				TriggerResult triggerResult = context.onProcessingTime(timer.timestamp);
 				if (triggerResult.isFire()) {
-					fire(context.window, contents);
+					emitWindowContents(context.window, contents);
 				}
 
-				if (triggerResult.isPurge() || (!windowAssigner.isEventTime() && isCleanupTime(context.window, timer.timestamp))) {
-					cleanup(context.window, windowState, mergingWindows);
+				if (!windowAssigner.isEventTime() && isCleanupTime(context.window, timer.timestamp)) {
+					clearAllState(context.window, windowState, mergingWindows);
+				} else if (triggerResult.isPurge()) {
+					windowState.clear();
 				}
 
+				if (mergingWindows != null) {
+					// need to make sure to update the merging state in state
+					mergingWindows.persist();
+				}
 			} else {
 				fire = false;
 			}
@@ -524,59 +543,35 @@ public class WindowOperator<K, IN, ACC, OUT, W extends Window>
 	}
 
 	/**
-	 * Cleans up the window state if the provided {@link TriggerResult} requires so, or if it
-	 * is time to do so (see {@link #isCleanupTime(Window, long)}). The caller must ensure that the
+	 * Drops all state for the given window and calls
+	 * {@link Trigger#clear(Window, Trigger.TriggerContext)}.
+	 *
+	 * <p>The caller must ensure that the
 	 * correct key is set in the state backend and the context object.
 	 */
-	private void cleanup(W window,
-						AppendingState<IN, ACC> windowState,
-						MergingWindowSet<W> mergingWindows) throws Exception {
+	private void clearAllState(
+			W window,
+			AppendingState<IN, ACC> windowState,
+			MergingWindowSet<W> mergingWindows) throws Exception {
 		windowState.clear();
+		context.clear();
 		if (mergingWindows != null) {
 			mergingWindows.retireWindow(window);
 		}
-		context.clear();
 	}
 
 	/**
-	 * Triggers the window computation if the provided {@link TriggerResult} requires so.
-	 * The caller must ensure that the correct key is set in the state backend and the context object.
+	 * Emits the contents of the given window using the {@link InternalWindowFunction}.
 	 */
 	@SuppressWarnings("unchecked")
-	private void fire(W window, ACC contents) throws Exception {
+	private void emitWindowContents(W window, ACC contents) throws Exception {
 		timestampedCollector.setAbsoluteTimestamp(window.maxTimestamp());
 		userFunction.apply(context.key, context.window, contents, timestampedCollector);
 	}
 
 	/**
-	 * Retrieves the {@link MergingWindowSet} for the currently active key.
-	 * The caller must ensure that the correct key is set in the state backend.
-	 */
-	@SuppressWarnings("unchecked")
-	protected MergingWindowSet<W> getMergingWindowSet() throws Exception {
-		MergingWindowSet<W> mergingWindows = mergingWindowsByKey.get((K) getStateBackend().getCurrentKey());
-		if (mergingWindows == null) {
-			// try to retrieve from state
-
-			TupleSerializer<Tuple2<W, W>> tupleSerializer = new TupleSerializer<>((Class) Tuple2.class, new TypeSerializer[] {windowSerializer, windowSerializer} );
-			ListStateDescriptor<Tuple2<W, W>> mergeStateDescriptor = new ListStateDescriptor<>("merging-window-set", tupleSerializer);
-			ListState<Tuple2<W, W>> mergeState = getStateBackend().getPartitionedState(VoidNamespace.INSTANCE, VoidNamespaceSerializer.INSTANCE, mergeStateDescriptor);
-
-			mergingWindows = new MergingWindowSet<>((MergingWindowAssigner<? super IN, W>) windowAssigner, mergeState);
-			mergeState.clear();
-
-			mergingWindowsByKey.put((K) getStateBackend().getCurrentKey(), mergingWindows);
-		}
-		return mergingWindows;
-	}
-
-	/**
-	 * Decides if a window is currently late or not, based on the current
-	 * watermark, i.e. the current event time, and the allowed lateness.
-	 * @param window
-	 * 					The collection of windows returned by the {@link WindowAssigner}.
-	 * @return The windows (among the {@code eligibleWindows}) for which the element should still be
-	 * 					considered when triggering.
+	 * Returns {@code true} if the watermark is after the end timestamp plus the allowed lateness
+	 * of the given window.
 	 */
 	protected boolean isLate(W window) {
 		return (windowAssigner.isEventTime() && (cleanupTime(window) <= currentWatermark));
@@ -588,11 +583,22 @@ public class WindowOperator<K, IN, ACC, OUT, W extends Window>
 	 * 					the window whose state to discard
 	 */
 	protected void registerCleanupTimer(W window) {
-		long cleanupTime = cleanupTime(window);
+
 		if (windowAssigner.isEventTime()) {
+			long cleanupTime = cleanupTime(window);
+			if (cleanupTime == Long.MAX_VALUE) {
+				// don't set a GC timer for "end of time"
+				return;
+			}
 			context.registerEventTimeTimer(cleanupTime);
 		} else {
-			context.registerProcessingTimeTimer(cleanupTime);
+			if (window.maxTimestamp() == Long.MAX_VALUE) {
+				// don't set a GC timer for "end of time"
+				return;
+			}
+
+			// for processing time we don't need allowed lateness
+			context.registerProcessingTimeTimer(window.maxTimestamp());
 		}
 	}
 
@@ -603,6 +609,10 @@ public class WindowOperator<K, IN, ACC, OUT, W extends Window>
 	 */
 	protected void deleteCleanupTimer(W window) {
 		long cleanupTime = cleanupTime(window);
+		if (cleanupTime == Long.MAX_VALUE) {
+			// no need to clean up because we didn't set one
+			return;
+		}
 		if (windowAssigner.isEventTime()) {
 			context.deleteEventTimeTimer(cleanupTime);
 		} else {
@@ -625,19 +635,15 @@ public class WindowOperator<K, IN, ACC, OUT, W extends Window>
 	}
 
 	/**
-	 * Decides if it is time to clean up the window state.
-	 * Clean up time for a window is:
-	 * 		<li> if it is event time, after the watermark passes the end of the window plus the user-specified allowed lateness
-	 * 		<li> if it is processing time, after the processing time at the node passes the end of the window.
-	 * 	@param window
-	 * 					the window to clean
-	 *  @param time
-	 *  				the current time (event or processing depending on the {@link WindowAssigner}
-	 *  @return {@code true} if it is time to clean up the window state, {@code false} otherwise.
+	 * Returns {@code true} if the given time is the cleanup time for the given window.
 	 */
 	protected final boolean isCleanupTime(W window, long time) {
-		long cleanupTime = cleanupTime(window);
-		return  cleanupTime == time;
+		if (windowAssigner.isEventTime()) {
+			long cleanupTime = cleanupTime(window);
+			return cleanupTime == time;
+		} else {
+			return window.maxTimestamp() == time;
+		}
 	}
 
 	/**
@@ -784,9 +790,9 @@ public class WindowOperator<K, IN, ACC, OUT, W extends Window>
 			return trigger.onEventTime(time, window, this);
 		}
 
-		public TriggerResult onMerge(Collection<W> mergedWindows) throws Exception {
+		public void onMerge(Collection<W> mergedWindows) throws Exception {
 			this.mergedWindows = mergedWindows;
-			return trigger.onMerge(window, this);
+			trigger.onMerge(window, this);
 		}
 
 		public void clear() throws Exception {
@@ -863,27 +869,13 @@ public class WindowOperator<K, IN, ACC, OUT, W extends Window>
 	@Override
 	@SuppressWarnings("unchecked")
 	public void snapshotState(FSDataOutputStream out, long checkpointId, long timestamp) throws Exception {
-
-		if (mergingWindowsByKey != null) {
-			TupleSerializer<Tuple2<W, W>> tupleSerializer = new TupleSerializer<>((Class) Tuple2.class, new TypeSerializer[] {windowSerializer, windowSerializer} );
-			ListStateDescriptor<Tuple2<W, W>> mergeStateDescriptor = new ListStateDescriptor<>("merging-window-set", tupleSerializer);
-			for (Map.Entry<K, MergingWindowSet<W>> key: mergingWindowsByKey.entrySet()) {
-				setKeyContext(key.getKey());
-				ListState<Tuple2<W, W>> mergeState = getStateBackend().getPartitionedState(VoidNamespace.INSTANCE, VoidNamespaceSerializer.INSTANCE, mergeStateDescriptor);
-				mergeState.clear();
-				key.getValue().persist(mergeState);
-			}
-		}
-
 		snapshotTimers(new DataOutputViewStreamWrapper(out));
-
 		super.snapshotState(out, checkpointId, timestamp);
 	}
 
 	@Override
 	public void restoreState(FSDataInputStream in) throws Exception {
 		restoreTimers(new DataInputViewStreamWrapper(in));
-
 		super.restoreState(in);
 	}
 
@@ -965,5 +957,15 @@ public class WindowOperator<K, IN, ACC, OUT, W extends Window>
 	@VisibleForTesting
 	public StateDescriptor<? extends AppendingState<IN, ACC>, ?> getStateDescriptor() {
 		return windowStateDescriptor;
+	}
+
+	@VisibleForTesting
+	public int numProcessingTimeTimers() {
+		return processingTimeTimers.size();
+	}
+
+	@VisibleForTesting
+	public int numEventTimeTimers() {
+		return watermarkTimers.size();
 	}
 }
